# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Star Wars API Documentation",
        "description": "API for accessing Star Wars data (People, Films, etc.)",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Star Wars API (Development)"
        }
    ],
    "paths": {
        "/analytics": {
            "get": {
                "operationId": "be3cd15e374cf0aca93d6eb3a3fe9941",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "average_request_timing": {
                                            "properties": {
                                                "average_duration_seconds": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0.002587865029151241
                                                },
                                                "average_duration_ms": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2.59
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "count_of_calls": {
                                            "properties": {
                                                "api_call_count": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "count_of_errors": {
                                            "properties": {
                                                "api_call_count_errors": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/star-wars/people": {
            "get": {
                "operationId": "3f10afef3eefadf2aa703b21139b904a",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the person to search for (e.g., 'luke')",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Luke Skywalker"
                                            },
                                            "uid": {
                                                "type": "string",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/star-wars/person": {
            "get": {
                "operationId": "1e35f1ffdc5e0f53592277721d8e6518",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "query",
                        "description": "Get one person by uid (e.g., '1')",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Luke Skywalker"
                                        },
                                        "birth_year": {
                                            "type": "string",
                                            "example": "19BBY"
                                        },
                                        "gender": {
                                            "type": "string",
                                            "example": "male"
                                        },
                                        "eye_color": {
                                            "type": "string",
                                            "example": "blue"
                                        },
                                        "hair_color": {
                                            "type": "string",
                                            "example": "blond"
                                        },
                                        "height": {
                                            "type": "string",
                                            "example": "172"
                                        },
                                        "mass": {
                                            "type": "string",
                                            "example": "77"
                                        },
                                        "movies": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "description": {
                                                        "type": "string",
                                                        "example": "A New Hope"
                                                    },
                                                    "uid": {
                                                        "type": "string",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/star-wars/movies": {
            "get": {
                "operationId": "6188c93c1585c6dcceacc977e51d3974",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the movie to search for (e.g., 'A New Hope')",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "A New Hope"
                                            },
                                            "uid": {
                                                "type": "string",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/star-wars/movie": {
            "get": {
                "operationId": "164e6078d6997a3bcf434f06986597e2",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "query",
                        "description": "Get one movie by uid (e.g., '1')",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Return of the Jedi"
                                        },
                                        "opening_crawl": {
                                            "type": "string",
                                            "example": "Synopsis"
                                        },
                                        "characters": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Luke Skywalker"
                                                    },
                                                    "uid": {
                                                        "type": "string",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}